refactor(tracing): add common tracing init to the utils library
bonus: Disable ansi escape characters
