+++ chart/templates/csi-node-daemonset.yaml
@@ -44,12 +44,13 @@ spec:
               fieldPath: status.podIP
         - name: RUST_BACKTRACE
           value: "1"
+        - name: RUST_LOG
+          value: "debug"
         args:
         - "--csi-socket=/csi/csi.sock"
         - "--node-name=$(MY_NODE_NAME)"
         - "--grpc-endpoint=$(MY_POD_IP):10199"{{ if .Values.csi.nvme.io_timeout_enabled }}
         - "--nvme-core-io-timeout={{ .Values.csi.nvme.io_timeout }}"{{ end }}
-        - "-v"
         command:
         - csi-node
         volumeMounts:
