--- common/src/opentelemetry.rs
@@ -1,36 +1,0 @@
-/// OpenTelemetry KeyVal for Processor Tags
-pub use opentelemetry::KeyValue;
-
-/// Parse KeyValues from structopt's cmdline arguments
-pub fn parse_key_value(source: &str) -> Result<KeyValue, String> {
-    match source.split_once('=') {
-        None => Err("Each element must be in the format: 'Key=Value'".to_string()),
-        Some((key, value)) => Ok(KeyValue::new(key.to_string(), value.to_string())),
-    }
-}
-
-/// Get Default Processor Tags
-/// ## Example:
-/// let _ = default_tracing_tags(git_version!(args = ["--abbrev=12", "--always"]),
-/// env!("CARGO_PKG_VERSION"));
-pub fn default_tracing_tags(git_commit: &str, cargo_version: &str) -> Vec<KeyValue> {
-    vec![
-        KeyValue::new("git.commit", git_commit.to_string()),
-        KeyValue::new("crate.version", cargo_version.to_string()),
-    ]
-}
-
-/// Name of the OTEL_BSP_MAX_EXPORT_BATCH_SIZE variable
-pub const OTEL_BSP_MAX_EXPORT_BATCH_SIZE_NAME: &str = "OTEL_BSP_MAX_EXPORT_BATCH_SIZE";
-/// The value of OTEL_BSP_MAX_EXPORT_BATCH_SIZE to be used with JAEGER
-pub const OTEL_BSP_MAX_EXPORT_BATCH_SIZE_JAEGER: &str = "64";
-/// Set the OTEL variables for a jaeger configuration
-pub fn set_jaeger_env() {
-    // if not set, default it to our jaeger value
-    if std::env::var(OTEL_BSP_MAX_EXPORT_BATCH_SIZE_NAME).is_err() {
-        std::env::set_var(
-            OTEL_BSP_MAX_EXPORT_BATCH_SIZE_NAME,
-            OTEL_BSP_MAX_EXPORT_BATCH_SIZE_JAEGER,
-        );
-    }
-}
