+++ control-plane/rest/service/src/main.rs
@@ -61,7 +61,7 @@ pub(crate) struct CliArgs {
     request_timeout: humantime::Duration,
 
     /// Add process service tags to the traces
-    #[structopt(short, long, env = "TRACING_TAGS", value_delimiter=",", parse(try_from_str = common_lib::opentelemetry::parse_key_value))]
+    #[structopt(short, long, env = "TRACING_TAGS", value_delimiter=",", parse(try_from_str = utils::tracing_telemetry::parse_key_value))]
     tracing_tags: Vec<KeyValue>,
 
     /// Don't use minimum timeouts for specific requests
@@ -90,45 +90,10 @@ use actix_web_opentelemetry::RequestTracing;
 use common_lib::{
     mbus_api,
     mbus_api::{BusClient, RequestMinTimeout, TimeoutOptions},
-    opentelemetry::default_tracing_tags,
 };
 use grpc::client::CoreClient;
 use http::Uri;
-use opentelemetry::{
-    global,
-    sdk::{propagation::TraceContextPropagator, trace::Tracer},
-    KeyValue,
-};
-
-fn init_tracing() -> Option<Tracer> {
-    if let Ok(filter) = tracing_subscriber::EnvFilter::try_from_default_env() {
-        tracing_subscriber::fmt().with_env_filter(filter).init();
-    } else {
-        tracing_subscriber::fmt().with_env_filter("info").init();
-    }
-    if let Some(agent) = CliArgs::args().jaeger {
-        let mut tracing_tags = CliArgs::args().tracing_tags;
-        tracing_tags.append(&mut default_tracing_tags(
-            utils::git_version(),
-            env!("CARGO_PKG_VERSION"),
-        ));
-        tracing_tags.dedup();
-        tracing::info!("Using the following tracing tags: {:?}", tracing_tags);
-        tracing::info!("Starting jaeger trace pipeline at {}...", agent);
-        // Start a new jaeger trace pipeline
-        global::set_text_map_propagator(TraceContextPropagator::new());
-        common_lib::opentelemetry::set_jaeger_env();
-        let tracer = opentelemetry_jaeger::new_pipeline()
-            .with_agent_endpoint(agent)
-            .with_service_name("rest-server")
-            .with_tags(tracing_tags)
-            .install_batch(opentelemetry::runtime::TokioCurrentThread)
-            .expect("Should be able to initialise the exporter");
-        Some(tracer)
-    } else {
-        None
-    }
-}
+use opentelemetry::{global, KeyValue};
 
 /// Extension trait for actix-web applications.
 pub trait OpenApiExt<T> {
@@ -211,9 +176,13 @@ fn get_jwk_path() -> Option<String> {
 
 #[actix_web::main]
 async fn main() -> anyhow::Result<()> {
-    // need to keep the jaeger pipeline tracer alive, if enabled
-    let _tracer = init_tracing();
     utils::print_package_info!();
+    let cli_args = CliArgs::args();
+    utils::tracing_telemetry::init_tracing(
+        "rest-server",
+        cli_args.tracing_tags.clone(),
+        cli_args.jaeger.clone(),
+    );
 
     let app = move || {
         App::new()
