+++ utils/utils-lib/Cargo.toml
@@ -7,3 +7,8 @@ edition = "2018"
 
 [dependencies]
 git-version = "0.3.5"
+tracing = "0.1.28"
+tracing-subscriber = "0.2.24"
+tracing-opentelemetry = "0.15.0"
+opentelemetry = { version = "0.16.0", features = ["rt-tokio-current-thread"] }
+opentelemetry-jaeger = { version = "0.15.0", features = ["rt-tokio-current-thread"] }
